Lag egen fil for aktivitet barnetilsyn
