+++ src/frontend/Sider/Behandling/Inngangsvilkår/Aktivitet/utilsBarnetilsyn.ts
@@ -3,11 +3,9 @@ import { Registeraktivitet } from '../../../../typer/registeraktivitet';
 import { dagensDato, førsteDagIMånedTreMånederForut } from '../../../../utils/dato';
 import { Periode } from '../../../../utils/periode';
 import { harTallverdi } from '../../../../utils/tall';
-import {
-    AktivitetBarnetilsyn,
-    AktivitetType,
-    AktivitetBarnetilsynFaktaOgSvar,
-} from '../typer/vilkårperiode/aktivitet';
+import { AktivitetType } from '../typer/vilkårperiode/aktivitet';
+import { AktivitetBarnetilsynFaktaOgSvar } from '../typer/vilkårperiode/aktivitetBarnetilsyn';
+import { AktivitetBarnetilsyn } from '../typer/vilkårperiode/aktivitetBarnetilsyn';
 import { SvarJaNei } from '../typer/vilkårperiode/vilkårperiode';
 import { BegrunnelseGrunner } from '../Vilkårperioder/Begrunnelse/utils';
